Update for 2.1.2.
2.1.2 was released on August 10th, 2014.

From the supervision mailing list:
```
version 2.1.2 of the runit package is available for through

 http://smarden.org/runit/

The sv program learns the LSB init script actions reload and
try-restart.  It also fixes two bugs that lead to wrong or badly
formatted output under some circumstances when reporting status of
services, and its documentation now properly states that on exit no
TERM signal is sent to the optional log service, but its standard
input is closed.
```